Add setListener function to respond to prompt_texts in ChatGPT translation
